moved Student Response code into ViewStudentResponse from ViewStudentApplications
